feat: clean up routes from main.go
