+++ backend/main.go
@@ -2,43 +2,18 @@ package main
 
 import (
 	"bytes"
-	"context"
-	"encoding/json"
 	"fmt"
+	"github.com/gin-gonic/gin"
+	"github.com/katamyra/kubestellarUI/routes"
+	"github.com/katamyra/kubestellarUI/wds/deployment"
+	"go.uber.org/zap"
 	"io"
 	"log"
 	"net/http"
 	"os"
 	"strings"
 	"time"
-
-	"github.com/gorilla/websocket"
-	"k8s.io/client-go/informers"
-
-	"github.com/gin-gonic/gin"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-
-	"github.com/katamyra/kubestellarUI/api"
-	nsresources "github.com/katamyra/kubestellarUI/namespace/resources"
-	"github.com/katamyra/kubestellarUI/wds"
-	"github.com/katamyra/kubestellarUI/wds/bp"
-	"github.com/katamyra/kubestellarUI/wds/deployment"
-	"github.com/katamyra/kubestellarUI/wds/resources"
-	"go.uber.org/zap"
 )
-
-type ContextInfo struct {
-	Name    string `json:"name"`
-	Cluster string `json:"cluster"`
-}
-
-type ManagedClusterInfo struct {
-	Name         string            `json:"name"`
-	Labels       map[string]string `json:"labels"`
-	CreationTime time.Time         `json:"creationTime"`
-	Context      string            `json:"context"`
-}
 
 func main() {
 
@@ -59,38 +34,11 @@ func main() {
 		}
 		c.Next()
 	})
-
-	router.GET("/api/clusters", func(c *gin.Context) {
-		contexts, clusters, currentContext, err, itsData := getKubeInfo()
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			return
-		}
-		c.JSON(http.StatusOK, gin.H{
-			"contexts":       contexts,
-			"clusters":       clusters,
-			"currentContext": currentContext,
-			"itsData":        itsData,
-		})
-	})
-
-	router.POST("/clusters/onboard", api.OnboardClusterHandler)
-	router.GET("/clusters/status", api.GetClusterStatusHandler)
-	router.POST("/clusters/import", api.ImportClusterHandler)
-
-	router.GET("/api/wds/workloads", func(c *gin.Context) {
-		workloads, err := deployment.GetWDSWorkloads()
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			return
-		}
-		c.JSON(http.StatusOK, workloads)
-	})
-
+	routes.SetupRoutes(router)
 	// New Log Endpoint
 	router.GET("/api/log", func(c *gin.Context) {
 		// Fetch Kubernetes Information
-		contexts, clusters, currentContext, _, itsData := getKubeInfo()
+		contexts, clusters, currentContext, _, itsData := routes.GetKubeInfo()
 
 		// Fetch WDS Workloads
 		workloads, err := deployment.GetWDSWorkloads()
@@ -140,66 +88,6 @@ func main() {
 		// Send the Log String
 		c.String(http.StatusOK, logBuilder.String())
 	})
-
-	// Route to CRUD deployment
-	router.POST("/api/wds/create", deployment.CreateDeployment)
-	router.POST("/api/wds/create/json", deployment.HandleCreateDeploymentJson)
-	router.PUT("/api/wds/update", deployment.UpdateDeployment)
-	router.DELETE("/api/wds/delete", deployment.DeleteDeployment)
-	router.GET("/api/wds/:name", deployment.GetDeploymentByName)
-	router.GET("/api/wds/status", deployment.GetDeploymentStatus)
-	// websocket
-	router.GET("/ws", func(ctx *gin.Context) {
-		deployment.HandleDeploymentLogs(ctx.Writer, ctx.Request)
-	})
-	// WATCH ALL DEPLOYMENT LOGS USING INFORMER
-	router.GET("/api/wds/logs", func(ctx *gin.Context) {
-		var upgrader = websocket.Upgrader{
-			CheckOrigin: func(r *http.Request) bool {
-				return true
-			},
-		}
-		var w = ctx.Writer
-		var r = ctx.Request
-		conn, err := upgrader.Upgrade(w, r, nil)
-		if err != nil {
-			log.Println("Failed to upgrade connection:", err)
-			ctx.JSON(http.StatusBadRequest, gin.H{"error": "Failed to upgrade to WebSocket"})
-			return
-		}
-		//defer conn.Close()
-
-		clientset, err := wds.GetClientSetKubeConfig()
-		if err != nil {
-			log.Println("Failed to get Kubernetes client:", err)
-			conn.WriteMessage(websocket.TextMessage, []byte("Error getting Kubernetes client"))
-			return
-		}
-		ch := make(chan struct{})
-		factory := informers.NewSharedInformerFactory(clientset, 10*time.Minute)
-		c := wds.NewController(clientset, factory.Apps().V1().Deployments(), conn)
-		factory.Start(ch)
-		go c.Run(ch)
-	})
-	// SERVICES
-	router.GET("/api/services/:namespace", resources.GetServiceList)
-	router.GET("/api/services/:namespace/:name", resources.GetServiceByServiceName)
-	router.POST("/api/services/create", resources.CreateService)
-	router.DELETE("/api/services/delete", resources.DeleteService)
-
-	// NAMESPACES
-	router.GET("/api/namespaces", nsresources.GetAllNamespaces)
-	router.GET("/api/namespaces/:name", nsresources.GetNamespaceDetails)
-	router.POST("/api/namespaces/create", nsresources.CreateNamespace)
-	router.PUT("/api/namespaces/update/:name", nsresources.UpdateNamespace)
-	router.DELETE("/api/namespaces/delete/:name", nsresources.DeleteNamespace)
-	router.GET("/ws/namespaces", nsresources.NamespaceWebSocketHandler)
-
-	router.POST("api/deploy", api.DeployHandler)
-	// ROUTES FOR BP
-	router.POST("/api/bp/create", bp.CreateBp)
-	router.DELETE("/api/bp/delete/:name", bp.DeleteBp)
-	router.DELETE("/api/bp/delete", bp.DeleteAllBp)
 
 	if err := router.Run(":4000"); err != nil {
 		log.Fatalf("Failed to start server: %v", err)
@@ -265,161 +153,6 @@ func ZapMiddleware() gin.HandlerFunc {
 			}
 		}
 	}
-}
-
-func getKubeInfo() ([]ContextInfo, []string, string, error, []ManagedClusterInfo) {
-	kubeconfig := os.Getenv("KUBECONFIG")
-	if kubeconfig == "" {
-		if home := homeDir(); home != "" {
-			kubeconfig = fmt.Sprintf("%s/.kube/config", home)
-		}
-		log.Printf("Using default kubeconfig path: %s", kubeconfig)
-	} else {
-		log.Printf("Using kubeconfig from environment: %s", kubeconfig)
-	}
-
-	config, err := clientcmd.LoadFromFile(kubeconfig)
-	if err != nil {
-		return nil, nil, "", err, nil
-	}
-
-	var contexts []ContextInfo
-	clusterSet := make(map[string]bool)
-	currentContext := config.CurrentContext
-
-	var managedClusters []ManagedClusterInfo
-
-	for contextName := range config.Contexts {
-		if strings.HasPrefix(contextName, "its") {
-			log.Printf("Processing ITS context: %s", contextName)
-
-			clientConfig := clientcmd.NewNonInteractiveClientConfig(
-				*config,
-				contextName,
-				&clientcmd.ConfigOverrides{},
-				clientcmd.NewDefaultClientConfigLoadingRules(),
-			)
-
-			restConfig, err := clientConfig.ClientConfig()
-			if err != nil {
-				log.Printf("Error creating REST config for context %s: %v", contextName, err)
-				continue
-			}
-
-			clientset, err := kubernetes.NewForConfig(restConfig)
-			if err != nil {
-				log.Printf("Error creating clientset for context %s: %v", contextName, err)
-				continue
-			}
-
-			clustersBytes, err := clientset.RESTClient().Get().
-				AbsPath("/apis/cluster.open-cluster-management.io/v1").
-				Resource("managedclusters").
-				DoRaw(context.TODO())
-
-			if err != nil {
-				log.Printf("Error fetching managed clusters from context %s: %v", contextName, err)
-				continue
-			}
-
-			var clusterList struct {
-				Items []struct {
-					Metadata struct {
-						Name              string            `json:"name"`
-						Labels            map[string]string `json:"labels"`
-						CreationTimestamp string            `json:"creationTimestamp"`
-					} `json:"metadata"`
-				} `json:"items"`
-			}
-
-			if err := json.Unmarshal(clustersBytes, &clusterList); err != nil {
-				log.Printf("Error unmarshaling clusters from context %s: %v", contextName, err)
-				continue
-			}
-
-			for _, item := range clusterList.Items {
-				creationTime, _ := time.Parse(time.RFC3339, item.Metadata.CreationTimestamp)
-				managedClusters = append(managedClusters, ManagedClusterInfo{
-					Name:         item.Metadata.Name,
-					Labels:       item.Metadata.Labels,
-					CreationTime: creationTime,
-					Context:      contextName,
-				})
-			}
-		}
-	}
-
-	// Process other kubeflex contexts (non-its)
-	for contextName, context := range config.Contexts {
-		if strings.HasSuffix(contextName, "-kubeflex") {
-			contexts = append(contexts, ContextInfo{
-				Name:    contextName,
-				Cluster: context.Cluster,
-			})
-			clusterSet[context.Cluster] = true
-		}
-	}
-
-	// Convert unique clusters to slice
-	var clusters []string
-	for clusterName := range clusterSet {
-		clusters = append(clusters, clusterName)
-	}
-
-	return contexts, clusters, currentContext, nil, managedClusters
-}
-
-// nolint:unused
-func getITSInfo() ([]ManagedClusterInfo, error) {
-	kubeconfig := os.Getenv("KUBECONFIG")
-	if kubeconfig == "" {
-		if home := homeDir(); home != "" {
-			kubeconfig = fmt.Sprintf("%s/.kube/config", home)
-		}
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
-	if err != nil {
-		return nil, err
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	clustersBytes, err := clientset.RESTClient().Get().
-		AbsPath("/apis/cluster.open-cluster-management.io/v1").
-		Resource("managedclusters").
-		DoRaw(context.TODO())
-	if err != nil {
-		return nil, fmt.Errorf("failed to get managed clusters: %w", err)
-	}
-
-	var clusterList struct {
-		Items []struct {
-			Metadata struct {
-				Name              string            `json:"name"`
-				Labels            map[string]string `json:"labels"`
-				CreationTimestamp string            `json:"creationTimestamp"`
-			} `json:"metadata"`
-		} `json:"items"`
-	}
-
-	if err := json.Unmarshal(clustersBytes, &clusterList); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal clusters: %w", err)
-	}
-
-	var managedClusters []ManagedClusterInfo
-	for _, item := range clusterList.Items {
-		managedClusters = append(managedClusters, ManagedClusterInfo{
-			Name:         item.Metadata.Name,
-			Labels:       item.Metadata.Labels,
-			CreationTime: time.Now().UTC(),
-		})
-	}
-
-	return managedClusters, nil
 }
 
 func homeDir() string {
