+++ backend/routes/cluster.go
@@ -1,0 +1,208 @@
+package routes
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"github.com/gin-gonic/gin"
+	"github.com/katamyra/kubestellarUI/api"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/clientcmd"
+	"log"
+	"net/http"
+	"os"
+	"strings"
+	"time"
+)
+
+type ManagedClusterInfo struct {
+	Name         string            `json:"name"`
+	Labels       map[string]string `json:"labels"`
+	CreationTime time.Time         `json:"creationTime"`
+	Context      string            `json:"context"`
+}
+type ContextInfo struct {
+	Name    string `json:"name"`
+	Cluster string `json:"cluster"`
+}
+
+func setupClusterRoutes(router *gin.Engine) {
+	router.GET("/api/clusters", func(c *gin.Context) {
+		contexts, clusters, currentContext, err, itsData := GetKubeInfo()
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
+		}
+		c.JSON(http.StatusOK, gin.H{
+			"contexts":       contexts,
+			"clusters":       clusters,
+			"currentContext": currentContext,
+			"itsData":        itsData,
+		})
+	})
+	router.POST("/clusters/onboard", api.OnboardClusterHandler)
+	router.GET("/clusters/status", api.GetClusterStatusHandler)
+	router.POST("/clusters/import", api.ImportClusterHandler)
+}
+
+func GetKubeInfo() ([]ContextInfo, []string, string, error, []ManagedClusterInfo) {
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		if home := homeDir(); home != "" {
+			kubeconfig = fmt.Sprintf("%s/.kube/config", home)
+		}
+		log.Printf("Using default kubeconfig path: %s", kubeconfig)
+	} else {
+		log.Printf("Using kubeconfig from environment: %s", kubeconfig)
+	}
+
+	config, err := clientcmd.LoadFromFile(kubeconfig)
+	if err != nil {
+		return nil, nil, "", err, nil
+	}
+
+	var contexts []ContextInfo
+	clusterSet := make(map[string]bool)
+	currentContext := config.CurrentContext
+
+	var managedClusters []ManagedClusterInfo
+
+	for contextName := range config.Contexts {
+		if strings.HasPrefix(contextName, "its") {
+			log.Printf("Processing ITS context: %s", contextName)
+
+			clientConfig := clientcmd.NewNonInteractiveClientConfig(
+				*config,
+				contextName,
+				&clientcmd.ConfigOverrides{},
+				clientcmd.NewDefaultClientConfigLoadingRules(),
+			)
+
+			restConfig, err := clientConfig.ClientConfig()
+			if err != nil {
+				log.Printf("Error creating REST config for context %s: %v", contextName, err)
+				continue
+			}
+
+			clientset, err := kubernetes.NewForConfig(restConfig)
+			if err != nil {
+				log.Printf("Error creating clientset for context %s: %v", contextName, err)
+				continue
+			}
+
+			clustersBytes, err := clientset.RESTClient().Get().
+				AbsPath("/apis/cluster.open-cluster-management.io/v1").
+				Resource("managedclusters").
+				DoRaw(context.TODO())
+
+			if err != nil {
+				log.Printf("Error fetching managed clusters from context %s: %v", contextName, err)
+				continue
+			}
+
+			var clusterList struct {
+				Items []struct {
+					Metadata struct {
+						Name              string            `json:"name"`
+						Labels            map[string]string `json:"labels"`
+						CreationTimestamp string            `json:"creationTimestamp"`
+					} `json:"metadata"`
+				} `json:"items"`
+			}
+
+			if err := json.Unmarshal(clustersBytes, &clusterList); err != nil {
+				log.Printf("Error unmarshaling clusters from context %s: %v", contextName, err)
+				continue
+			}
+
+			for _, item := range clusterList.Items {
+				creationTime, _ := time.Parse(time.RFC3339, item.Metadata.CreationTimestamp)
+				managedClusters = append(managedClusters, ManagedClusterInfo{
+					Name:         item.Metadata.Name,
+					Labels:       item.Metadata.Labels,
+					CreationTime: creationTime,
+					Context:      contextName,
+				})
+			}
+		}
+	}
+
+	// Process other kubeflex contexts (non-its)
+	for contextName, context := range config.Contexts {
+		if strings.HasSuffix(contextName, "-kubeflex") {
+			contexts = append(contexts, ContextInfo{
+				Name:    contextName,
+				Cluster: context.Cluster,
+			})
+			clusterSet[context.Cluster] = true
+		}
+	}
+
+	// Convert unique clusters to slice
+	var clusters []string
+	for clusterName := range clusterSet {
+		clusters = append(clusters, clusterName)
+	}
+
+	return contexts, clusters, currentContext, nil, managedClusters
+}
+
+func homeDir() string {
+	if h := os.Getenv("HOME"); h != "" {
+		return h
+	}
+	return os.Getenv("USERPROFILE") // windows
+}
+
+// nolint:unused
+func getITSInfo() ([]ManagedClusterInfo, error) {
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		if home := homeDir(); home != "" {
+			kubeconfig = fmt.Sprintf("%s/.kube/config", home)
+		}
+	}
+
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, err
+	}
+
+	clustersBytes, err := clientset.RESTClient().Get().
+		AbsPath("/apis/cluster.open-cluster-management.io/v1").
+		Resource("managedclusters").
+		DoRaw(context.TODO())
+	if err != nil {
+		return nil, fmt.Errorf("failed to get managed clusters: %w", err)
+	}
+
+	var clusterList struct {
+		Items []struct {
+			Metadata struct {
+				Name              string            `json:"name"`
+				Labels            map[string]string `json:"labels"`
+				CreationTimestamp string            `json:"creationTimestamp"`
+			} `json:"metadata"`
+		} `json:"items"`
+	}
+
+	if err := json.Unmarshal(clustersBytes, &clusterList); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal clusters: %w", err)
+	}
+
+	var managedClusters []ManagedClusterInfo
+	for _, item := range clusterList.Items {
+		managedClusters = append(managedClusters, ManagedClusterInfo{
+			Name:         item.Metadata.Name,
+			Labels:       item.Metadata.Labels,
+			CreationTime: time.Now().UTC(),
+		})
+	}
+
+	return managedClusters, nil
+}
